requirements.txt
textual>=0.85.0
aiohttp>=3.9.0
websockets>=12.0
aiosqlite>=0.19.0
pyyaml>=6.0
py-clob-client>=0.15.0
python-dotenv>=1.0.0
textual-plotext>=0.2.1
